requirements.txt
asgiref==3.7.2
Django==5.0.3
filelock==3.13.1
fsspec==2024.3.0
Jinja2==3.1.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
pillow==10.2.0
sqlparse==0.4.4
sympy==1.12
torch==2.2.1
torchvision==0.17.1
typing_extensions==4.10.0
tzdata==2024.1